fix: get blank page when filled threshold number is too large
Also apply right formating

risk: low
JIRA: F1-1104
